requirements.txt
streamlit
bcrypt
easyocr
google-genai
opencv-python
pandas
pillow
python-dotenv
pymupdf